dollars: Properly deal with $ in drivers
These are rare.  But they ought not to cause things to go haywire (as
I could previously achieve with the test case which follows shortly).

This would have been an awkward breaking change if we had released the
cross-crate machinery.  I think we should make some space for future
things of this kind.  That will come in a moment.