Abolish Spanned for paste::Item
Spanned can be confusing: do we mean template span, or driver span?

When pasting, we want to make sure we always use template kw span for
the output; that's in Items.

We might want a span for error reporting, of course.  But it turns out
that the only errors we might emit are for the nontrivial entries.
So we can get rid Item::Plain's span and the arguemnts that feed it.