Resolve a mysterious TODO
Here, these *are* in [ ].  I think this was referring to the input
syntax in define_derive_adhoc! and derive_adhoc!, where options
precede the `:` or `=`.

I think it's best to leave the syntax cleaner there, even though this
somewhat restricts the syntax of options (since they need to be
delimitable by both `;` and `=`).