Within ${paste }, reject values that aren't string literals
That booleans and integers were supported wasn't documented.

Integers aren't great because the text in the source might be changed
due to parsing and reformatting as an integer, so that the pasted
identifier is not what was expected.

Allowing Verbatim was just wrong.

Then, making booleans an exception seems too odd.