Introduce IdentAny helper type
And use it everywhere.  Enforce that keywords using keyword! are
IdentAny.

This is slightly more ergonomic at the call sites, and also ensures
that we don't have any keyword! that don't accept Rust keywords.

And, we're going to have call site where we want to use
syn::parse_str, which will *requires* this wrapper type, so we might
as well make it a proper thing and use it throughout.