Rework ${Xmeta} and ${Xmeta as ...} handling
No longer call append_meta_value, which is the thing that allows each
context to handle the default case (no `as`) separately.  Instead, we
now always treat a lack of `as` the same way: we parse the value as
tokens.

Provide `as tokens` to explicitly request the default behaviour.

Replace `as lit` with `as str`.

Now values for `#[adhoc]` are always expected to be string literals.
(Although, this is only checked if the template actually tries to use
the value.)

Another consequence is that `${paste ${Xmeta(thing)}}` isn't allowed
without the `... as`, since that means `as tokens` and you can't paste
with arbitrary tokens.  (The new arrangements correctly required us to
implement it this way: we need an `O::NotInPaste` to pass to
`append_tokens`, so `SubstMetaAs::Unspecified` has to contain a
`O::NotInPaste`.)

Document the new situation and make the necessary adjustment to a
test.

Fixes #21