case changing: Fix check-keywords-documented's handling of ref table
Fix the following bug:

check-keywords-documented looks at all the columns except the heck
one, and expects them all to be aliases.  That's not consistent with
the table's rubric, which doesn't say that the "Example of results"
was also supposed to be an alias.

This doesn't actually matter right now because those aliases are
duplicated in the earlier columns, but it's a latent bug which is
about to become a real bug.