example checking: possibilities::Limit::parse: new Others handling
Centralise the handling of the list of previous "for" clause, for
handling of "for others".

This moves the appending of the current clause to the list, into the
Limit::parse.  Now, that doesn't happen for Others, so that after
on "for others", the list is reset and a second batch can begin.

This fixes a latent bug, where multiple sets of "for this, for that,
for others" appeared.  It also changes the behaviour with consecutive
"for others": the second one now must match *everywhere*, not
nowhere (ie, it is no longer ignored).