Fix handling of Xmeta to not search the context tree
Previously $fmeta would be tolerated outside any field repetition: it
would search the whole struct for any field with that attribute.

This is very surprising and not documented.  I don't know why I
thought it was a good idea.

If we want such a thing it should be a   any_of(fields, ...
condition, or something.

This leaves the code a little flabby.  We'll sort that out later.