doc: reference attr examples: Fix attribute visibility
As just noted in the reference doc, attributes aren't visible to a
derive macro unless they come *after* the line with the derive.
This is done by the compiler before it gives the macro the TokenStream.

Because we're not actually using the parsing and processing from the
real compiler, our tests got this wrong.  Fix that by manually
simulating the compiler behaviour.

And, finally, we can fix the docs to be accurate.

When we fix the teststo properly hide the attribute containing
the #[derive(Adhoc), and everything before it, we'll

FILTER OUT ALL ATTRIBUTES (WHOLE #[...]) UP TO AND INCLUDING THE ONE
WITH #[DERIVE(ADHOC)]

F

squash! MUST